added jenkins pipeline

+++ sci-calc-backend/app.py
@@ -7,29 +7,29 @@ CORS(app)
 @app.route('/sqrt', methods=['POST'])
 def calculate_square_root():
     data = request.get_json()
-    num = int(data['number'])
+    num = float(data['number'])
     result = math.sqrt(num)
     return jsonify({'result': result})
 
 @app.route('/factorial', methods=['POST'])
 def calculate_factorial():
     data = request.get_json()
-    num = int(data['number'])
+    num = float(data['number'])
     result = math.factorial(num)
     return jsonify({'result': result})
 
 @app.route('/ln', methods=['POST'])
 def calculate_ln():
     data = request.get_json()
-    num = int(data['number'])
+    num = float(data['number'])
     result = math.log(num)
     return jsonify({'result': result})
 
 @app.route('/power', methods=['POST'])
 def calculate_power():
     data = request.get_json()
-    num = int(data['number'])
-    power = int(data['power'])
+    num = float(data['number'])
+    power = float(data['power'])
     result = math.pow(num, power)
     return jsonify({'result': result})
